extracted codeActionProvider class to sepparate util file
